only admin can edit/delete

+++ resources/views/pages/book.blade.php
@@ -25,13 +25,16 @@
                 <h2><strong>Description</strong></h2>
                 <p>{{$book->description}}</p>
                 <hr>
-                <a href="/iandwe/public/admin/{{$book->book_id}}/edit" class="btn btn-primary btn-lg">Edit</a><br/>
-                <br>
-                {!! Form::model($book, array('route' => array('admin.destroy', $book->book_id), 'method' => 'DELETE')) !!}
-                    {{Form::submit('Delete', ['class' => 'btn btn-lg btn-danger'] )}}
-                {!! Form::close() !!}
 
-                <br>
+                @if(session('session_admin_login')!= null)
+                    <a href="/iandwe/public/admin/{{$book->book_id}}/edit" class="btn btn-primary btn-lg">Edit</a><br/>
+                    <br>
+                    {!! Form::model($book, array('route' => array('admin.destroy', $book->book_id), 'method' => 'DELETE')) !!}
+                        {{Form::submit('Delete', ['class' => 'btn btn-lg btn-danger'] )}}
+                    {!! Form::close() !!}
+
+                    <br>
+                @endif
                 
 
             </div>
